requirements.txt
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
google-ads==25.1.0
google-api-core==2.23.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
oauthlib==3.2.2
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
urllib3==2.2.3
Werkzeug==3.1.3
zipp==3.21.0
